requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
email-validator==2.2.0
